# spec/api.yaml
...
openapi: 3.0.1
info:
  title: Corona Network
  description: |
    Restful API for the Corona Network decentralized social network.

    *The Corona Network API is not a globally accessible service, rather a server running locally on your device. The base URL is not a publicly routed domain, but rather a local one existing only on your device (and even on your device only within a process running [go-coronanet](https://github.com/coronanet/go-coronanet)).*
  version: 0.0.5
...
servers:
  - url: https://corona-network/
...
paths:
...
  /profile:
    post:
      summary: Create a new local user
      tags:
        - Profile
      responses:
        409:
          description: Local user already exists
        200:
          description: Successfuly created new user
    get:
      summary: Retrieves the local user's profile
      tags:
        - Profile
      responses:
        404:
          description: Local user doesn't exist
        200:
          $ref: '#/components/responses/Profile'
    put:
      summary: Updates the local user's profile
      tags:
        - Profile
      requestBody:
        description: New profile of the current user
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Profile'
      responses:
        400:
          description: Provided profile is invalid
        403:
          description: Local user doesn't exist
        200:
          description: User profile updated
    delete:
      summary: Deletes the local user (nukes all data)
      tags:
        - Profile
      responses:
        200:
          description: Successfully deleted user
...
  /pairing:
    post:
      summary: Creates a pairing session for contact establishment
      tags:
        - Contacts
      responses:
        403:
          description: Cannot pair while offline or without profile
        409:
          description: Pairing session already in progress
        200:
          description: Successfully created pairing session
          content:
            application/json:
              schema:
                type: string
                description: Temporary pairing secret
    get:
      summary: Waits for a pairing session to complete
      tags:
        - Contacts
      responses:
        403:
          description: No pairing session in progress
        409:
          description: Remote contact already paired
        200:
          description: Successfully established session
          content:
            application/json:
              schema:
                type: string
                description: Contact ID of the paired user
    put:
      summary: Joins a pairing session for contact establishment
      tags:
        - Contacts
      requestBody:
        description: Optional description in *Markdown*
        required: true
        content:
          application/json:
            schema:
              type: string
              description: Temporary pairing secret
      responses:
        400:
          description: Provided pairing secret is invalid
        403:
          description: Cannot pair while offline or without profile
        409:
          description: Remote contact already paired
        200:
          description: Successfully established session
          content:
            application/json:
              schema:
                type: string
                description: Contact ID of the paired user
  /contacts:
    get:
      summary: Lists all contacts of the local user
      tags:
        - Contacts
      responses:
        403:
          description: Local user doesn't exist
        200:
          description: Returns a list of contact IDs
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
...
components:
  schemas:
    Profile:
      type: object
      properties:
        name:
          type: string
          description: Full name of the user
...
  responses:
    Profile:
      description: Returns requested profile infos
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Profile'
...